Refactor newExpenseForm component into modular structure
- Deleted the original newExpenseForm.vue and split its functionality into three components: formContent.vue, formFooter.vue, and newExpenseForm.vue.
- Updated dashboard.vue to import the new newExpenseForm structure, ensuring proper component organization and improved maintainability.
- Enhanced the form structure for better readability and separation of concerns.
